feature(prod): check database.yml + run migrate:reset

+++ db/schema.rb
@@ -18,7 +18,7 @@ ActiveRecord::Schema.define(version: 2023_07_23_173430) do
   create_table "answers", force: :cascade do |t|
     t.text "body"
     t.boolean "correct", default: false, null: false
-    t.integer "question_id", null: false
+    t.bigint "question_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["question_id"], name: "index_answers_on_question_id"
@@ -33,7 +33,7 @@ ActiveRecord::Schema.define(version: 2023_07_23_173430) do
   create_table "gists", force: :cascade do |t|
     t.string "github_url"
     t.string "author_email"
-    t.integer "question_id", null: false
+    t.bigint "question_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["question_id"], name: "index_gists_on_question_id"
@@ -41,15 +41,15 @@ ActiveRecord::Schema.define(version: 2023_07_23_173430) do
 
   create_table "questions", force: :cascade do |t|
     t.text "body"
-    t.integer "test_id", null: false
+    t.bigint "test_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["test_id"], name: "index_questions_on_test_id"
   end
 
   create_table "test_passages", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.integer "test_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "test_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.integer "current_question_id"
@@ -61,8 +61,8 @@ ActiveRecord::Schema.define(version: 2023_07_23_173430) do
   create_table "tests", force: :cascade do |t|
     t.string "title", null: false
     t.integer "level", default: 1, null: false
-    t.integer "category_id", null: false
-    t.integer "author_id"
+    t.bigint "category_id", null: false
+    t.bigint "author_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["author_id"], name: "index_tests_on_author_id"
